Adding recent10me as a command
I'll leave this here in case you want to add; if not, simply deny the PR.

Functions the same as recent5me but it gives 10 plays instead of 5. Members will have a 10-minute cooldown for this command to avoid spam, while moderators will bypass this restriction. This command is restricted to #bot-ground only.